Docs update - using environment-dependent settings in configuration
It might not be the best solution, take it just as a suggestion for a missing chapter in docs. It took me a lot of time to figure this out and sincerely, I'm looking forward to seeing the correct solution for this. My code is based on http://p373.net/2014/01/25/how-to-create-environment-based-builds-with-grunt/, which didn't work for me entirely.